feat: refactor openfeature python
