+++ clients/openfeature/python/.gitignore
@@ -1,2 +1,1 @@
 __pycache__
-libsuperposition_core.dylib
