+++ clients/openfeature/python/pyproject.toml
@@ -1,15 +1,24 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
 [project]
-name = "openfeature_provider_python"
+name = "openfeature_superposition_provider"
 version = "0.0.1"
-description = "openfeature_provider_python"
+description = "openfeature_superposition_provider"
 readme = "README.md"
 requires-python = ">=3.12"
-keywords = ["openfeature", "openfeature_provider_python"]
+keywords = ["openfeature", "openfeature_superposition_provider"]
 license = {text = "Apache-2.0"}
 
 dependencies = [
     "smithy_core",
-    "superposition_python_sdk",
-    "json",
-    "asyncio"
+    "superposition_python_sdk @ git+https://github.com/juspay/superposition.git@main#subdirectory=clients/generated/smithy/python"
 ]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["openfeature_superposition_provider"]
+
+[tool.setuptools.package-data]
+"openfeature_superposition_provider" = ["libsuperposition_core.dylib"]
